Task #199 atribute fixes to integrate with DB

+++ src/main/java/com/EcoMentor_backend/EcoMentor/Certificate/useCases/dto/BorrarOfficialCertificateDTO.java
@@ -34,7 +34,7 @@ public class BorrarOfficialCertificateDTO {
     private String climateZone;
 
 
-    private Integer cadastreMeters;
+    private float cadastreMeters;
 
     private Year buildingYear;
