+++ src/main/java/com/EcoMentor_backend/EcoMentor/Certificate/useCases/dto/OfficialCertificateWFE.java
@@ -28,7 +28,7 @@ public class OfficialCertificateWFE {
     private String climateZone;
 
 
-    private Integer cadastreMeters;
+    private float cadastreMeters;
 
     private Year buildingYear;
